PeerSwitching optimization and fixes
